+++ src/main/java/org/kumoricon/registration/model/attendee/AttendeeRepository.java
@@ -166,14 +166,15 @@ public class AttendeeRepository {
                             emergency_contact_full_name, emergency_contact_phone, fan_name, first_name, last_name,
                             legal_first_name, legal_last_name, name_is_legal_name, preferred_pronoun, paid,
                             paid_amount, parent_form_received, parent_full_name, parent_is_emergency_contact,
-                            parent_phone, phone_number, pre_registered, zip, order_id, membership_revoked, accessibility_sticker) VALUES
+                            parent_phone, phone_number, pre_registered, zip, order_id, membership_revoked, 
+                            accessibility_sticker, last_modified) VALUES
                             (:badgeId, :badgeNumber, :badgePrePrinted, :badgePrinted,
                              :birthDate, :checkInTime, :checkedIn, :compedBadge, :country, :email,
                              :emergencyContactFullName, :emergencyContactPhone, :fanName, :firstName, :lastName,
                              :legalFirstName, :legalLastName, :nameIsLegalName, :preferredPronoun, :paid,
                              :paidAmount, :parentFormReceived, :parentFullName, :parentIsEmergencyContact,
                              :parentPhone, :phoneNumber, :preRegistered, :zip, :orderId, :membershipRevoked,
-                             :accessibilitySticker)
+                             :accessibilitySticker, now())
                             """, params);
         } else {
             jdbcTemplate.update("""
@@ -188,7 +189,7 @@ public class AttendeeRepository {
                             parent_full_name=:parentFullName, parent_is_emergency_contact=:parentIsEmergencyContact,
                             parent_phone=:parentPhone, phone_number=:phoneNumber,
                             zip=:zip, order_id=:orderId, membership_revoked=:membershipRevoked,
-                            accessibility_sticker=:accessibilitySticker WHERE id = :id
+                            accessibility_sticker=:accessibilitySticker, last_modified=now() WHERE id = :id
                             """, params);
         }
     }
@@ -259,14 +260,14 @@ public class AttendeeRepository {
                             legal_first_name, legal_last_name, name_is_legal_name, preferred_pronoun, paid,
                             paid_amount, parent_form_received, parent_full_name, parent_is_emergency_contact,
                             parent_phone, phone_number, pre_registered, zip, order_id, membership_revoked,
-                            accessibility_sticker) VALUES
+                            accessibility_sticker, last_modified) VALUES
                             (:badgeId, :badgeNumber, :badgePrePrinted, :badgePrinted,
                              :birthDate, :checkInTime, :checkedIn, :compedBadge, :country, :email,
                              :emergencyContactFullName, :emergencyContactPhone, :fanName, :firstName, :lastName,
                              :legalFirstName, :legalLastName, :nameIsLegalName, :preferredPronoun, :paid,
                              :paidAmount, :parentFormReceived, :parentFullName, :parentIsEmergencyContact,
                              :parentPhone, :phoneNumber, :preRegistered, :zip, :orderId, :membershipRevoked,
-                             :accessibilitySticker)
+                             :accessibilitySticker, now())
                             """, toInsertParams);
         }
 
@@ -290,7 +291,7 @@ public class AttendeeRepository {
                             parent_full_name=:parentFullName, parent_is_emergency_contact=:parentIsEmergencyContact,
                             parent_phone=:parentPhone, phone_number=:phoneNumber,
                             zip=:zip, order_id=:orderId, membership_revoked=:membershipRevoked,
-                            accessibility_sticker=:accessibilitySticker WHERE id = :id
+                            accessibility_sticker=:accessibilitySticker, last_modified=now() WHERE id = :id
                             """, toUpdateParams);
         }
     }
@@ -298,7 +299,7 @@ public class AttendeeRepository {
     @Transactional
     public void setParentFormReceived(Integer attendeeId, Integer orderId, Boolean formReceived) {
         final String SQL = """
-                UPDATE attendees SET parent_form_received = :parentFormReceived
+                UPDATE attendees SET parent_form_received = :parentFormReceived, last_modified=now()
                 WHERE id=:attendeeId AND order_id = :orderId
                 """;
         SqlParameterSource params = new MapSqlParameterSource("attendeeId", attendeeId)
@@ -348,6 +349,7 @@ public class AttendeeRepository {
             a.setOrderId(rs.getInt("order_id"));
             a.setMembershipRevoked(rs.getBoolean("membership_revoked"));
             a.setAccessibilitySticker(rs.getBoolean("accessibility_sticker"));
+            a.setLastModified(rs.getObject("last_modified", OffsetDateTime.class));
             return a;
         }
     }